使用ywjno.dev提交的代码,fix issue 485
git-svn-id: http://nutz.googlecode.com/svn/trunk@2115 423f10f2-e3a4-11dd-a6ea-a32d6b26a33d

+++ src/org/nutz/el/val/ArrayElValue.java
@@ -1,6 +1,7 @@
 package org.nutz.el.val;
 
 import java.lang.reflect.Array;
+import java.math.BigDecimal;
 
 import org.nutz.el.El;
 import org.nutz.el.ElException;
@@ -14,8 +15,8 @@ public class ArrayElValue extends PojoElValue<Object> {
 
 	public ElValue getProperty(ElValue val) {
 		Object v = val.get();
-		if (v instanceof Integer) {
-			return El.wrap(Array.get(obj, ((Integer) v).intValue()));
+		if (v instanceof BigDecimal) {
+			return El.wrap(Array.get(obj, ((BigDecimal) v).intValue()));
 		} else if ("length".equals(val.getString())) {
 			return El.wrap(Array.getLength(obj));
 		}
